munged: Compare decode time as uint32_t
dec_validate_time() compared "m->time1" (uint32_t) against "tmin"
and "tmax" (time_t).  On ILP32, time_t is 32-bit signed, so GCC flags
this under "-Wsign-compare".

Cast "tmin" and "tmax" to uint32_t.  Both are computed from "m->time0"
(uint32_t) in 32-bit unsigned arithmetic, so both fit a uint32_t.
The cast changes no result on ILP32 or LP64.

Casting "m->time1" to time_t instead would break on a 32-bit signed
time_t.  A timestamp at or past 2038 becomes negative, and a valid
credential is rejected as rewound.

+++ src/munged/dec.c
@@ -1033,11 +1033,15 @@ dec_validate_time (munge_cred_t c)
     tmax = m->time0 + m->ttl;
     /*
      *  Check the decode time against the allowable min & max.
+     *
+     *  tmin and tmax are computed from m->time0 (uint32_t) in 32-bit
+     *    unsigned arithmetic, so both fit a uint32_t and the casts are
+     *    value-preserving.
      */
-    if (m->time1 < tmin) {
+    if (m->time1 < (uint32_t) tmin) {
         return m_msg_set_err (m, EMUNGE_CRED_REWOUND, NULL);
     }
-    if (m->time1 > tmax) {
+    if (m->time1 > (uint32_t) tmax) {
         return m_msg_set_err (m, EMUNGE_CRED_EXPIRED, NULL);
     }
     return 0;
